+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Features:
 
   - add `OOD_USER_ENV` to specify CGI env var used for authenticated username
+  - deprecated the `/nginx/start` url option
 
 ## 0.2.0 (2017-01-30)